iox-#1032 Configurable handling
