--- iceoryx_hoofs/include/iceoryx_hoofs/error_handling_2/platform/handling.hpp
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "error_levels.hpp"
-#include "report.hpp"
-
-namespace eh
-{
-} // namespace eh
